Move everything to fix circular import
