+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from "./result";
-export * from "./initializers";
 export * from "./errors";
-export * from "./functions";
+export * from "./validator";
